rename slider-panel to tera-slider-panel
