+++ packages/client/hmi-client/src/components/widgets/tera-slider-panel.vue
