+++ packages/client/hmi-client/src/page/data-explorer/DataExplorer.vue
@@ -1,6 +1,6 @@
 <template>
 	<main>
-		<slider-panel
+		<tera-slider-panel
 			content-width="240px"
 			direction="left"
 			header="Facets"
@@ -15,7 +15,7 @@
 					:result-type="resourceType"
 				/>
 			</template>
-		</slider-panel>
+		</tera-slider-panel>
 		<div class="results-content">
 			<div class="secondary-header">
 				<span class="p-buttonset">
@@ -63,7 +63,7 @@
 			:search-term="searchTerm"
 			@toggle-data-item-selected="toggleDataItemSelected"
 		/>
-		<slider-panel
+		<tera-slider-panel
 			class="resources-slider"
 			:content-width="sliderWidth"
 			direction="right"
@@ -103,13 +103,13 @@
 					@find-similar-content="onFindSimilarContent"
 				/>
 			</template>
-		</slider-panel>
+		</tera-slider-panel>
 	</main>
 </template>
 
 <script setup lang="ts">
 import { computed, onMounted, onUnmounted, ref, watch } from 'vue';
-import SliderPanel from '@/components/widgets/slider-panel.vue';
+import TeraSliderPanel from '@/components/widgets/tera-slider-panel.vue';
 import { fetchData, getXDDSets } from '@/services/data';
 import {
 	Facets,
